some fixes
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@426 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/Misc.cpp
@@ -16,6 +16,7 @@
 
 #include "defs.h"
 #include "LieroX.h"
+#include "GfxPrimitives.h"
 
 
 
@@ -627,7 +628,7 @@ short stringcasecmp(const std::string& s1, const std::string& s2) {
 			// not at end of s1
 			return 1; // s1 > s2
 
-		dif = tolower(*p1) - tolower(*p2);
+		dif = (short)tolower(*p1) - (short)tolower(*p2);
 		if(dif != 0) return dif; // dif > 0  <=>  s1 > s2
 
 		p1++; p2++;
@@ -652,8 +653,7 @@ std::vector<std::string>& explode(const std::string& str, const std::string& del
 }
 
 // reads maxlen-1 chars from fp
-std::string freadstr(FILE *fp, size_t maxlen)
-{
+std::string freadstr(FILE *fp, size_t maxlen) {
 	if (!fp) return "";
 
 	static std::string result;
